+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roomplayer/YoutubeRoomPlayer.kt
@@ -3,16 +3,11 @@ package com.digginroom.digginroom.feature.room.customview.roomplayer
 import android.content.Context
 import android.view.LayoutInflater
 import android.webkit.JavascriptInterface
-import android.webkit.WebResourceRequest
-import android.webkit.WebResourceResponse
 import android.webkit.WebView
-import android.webkit.WebViewClient
 import android.widget.FrameLayout
 import com.digginroom.digginroom.databinding.ItemRoomInfoBinding
 import com.digginroom.digginroom.feature.room.roominfo.RoomInfoUiState
 import com.digginroom.digginroom.model.mapper.ScrapCountFormatter
-import java.io.ByteArrayInputStream
-import java.io.InputStream
 
 class YoutubeRoomPlayer(
     context: Context,
@@ -67,158 +62,20 @@ class YoutubeRoomPlayer(
     }
 
     private fun initYoutubePlayer() {
-        val iframe = """
-            <!DOCTYPE html>
-            <html lang="en">
-              <script>
-                var tag = document.createElement('script');
-
-                tag.src = "https://www.youtube.com/iframe_api";
-                var firstScriptTag = document.getElementsByTagName('script')[0];
-                firstScriptTag.parentNode.insertBefore(tag, firstScriptTag);
-
-                var isPlayerLoaded = false
-                var player
-
-                function play() {
-                    if (!isPlayerLoaded) return
-                    player.playVideo()
-                    player.unMute()
-                }
-                
-                function pause() {
-                    if (!isPlayerLoaded) return
-                    player.mute()
-                }
-                
-                function mute() {
-                    if (!isPlayerLoaded) return
-                    player.mute()
-                }
-
-                function unMute() {
-                    if (!isPlayerLoaded) return
-                    player.unMute()
-                }
-                
-                function vh(percent) {
-                    var h = Math.max(document.documentElement.clientHeight, window.innerHeight || 0);
-                    return (percent * h) / 100;
-                }
-        
-                function vw(percent) {
-                    var w = Math.max(document.documentElement.clientWidth, window.innerWidth || 0);
-                    return (percent * w) / 100;
-                }
-
-                function navigate(videoId) {
-                    if (!isPlayerLoaded) return
-                    
-                    const ele = document.getElementsByClassName("player-container")[0]
-                    ele.style.position = "absolute"
-                    const url = "https://www.youtube.com/oembed?url=http%3A//www.youtube.com/watch?v%3D" + videoId + "&format=json"
-                    fetch(url).then(function (response) {
-                        return response.json();
-                    }).then(function (data) {
-                        const whRatio = data["width"] / data["height"]
-                        ele.style.width = vh(whRatio * 100).toString() + "px"
-                        ele.style.left = "50%"
-                        ele.style.transform = "translate(-50%, 0%)"
-                    }).catch(error => {
-                        ele.style.width = "0%"
-                        ele.style.left = "0%"
-                        ele.style.right = "0%"
-                    });
-                    player.loadVideoById(videoId, 0, 'highres')
-                }
-
-                function onYouTubeIframeAPIReady() {
-                    player = new YT.Player('player', {
-                        events: {
-                            'onReady': onPlayerReady,
-                            'onStateChange': onPlayerStateChange
-                        },
-                        videoId: '',
-                        playerVars: {
-                            autoplay: 1,
-                            controls: 0,
-                            disablekb: 1,
-                            fs: 0,
-                            loop: 1,
-                            modestbranding: 1,
-                            rel: 0,
-                            showinfo: 0,
-                            mute: 1,
-                            autohide: 1,
-                        },
-                    })
-                }
-
-                function onPlayerReady() {
-                    Player.onLoaded()
-                    isPlayerLoaded = true
-                }
-                
-                function onPlayerStateChange(event) {
-                    if (event.data == 0)
-                        player.playVideo()
-                    if (event.data == 1)
-                        Player.onPlay()
-                }
-              </script>
-              <style>
-                html,
-                body,
-                #player,
-                .player-container {
-                  height: 100%;
-                  width: 100%;
-                  margin: 0;
-                  overflow: hidden;
-                }
-            
-                .player-container {
-                  position: absolute;
-                  top: -60px;
-                  height: calc(100% + 120px);
-                }
-              </style>
-              <head>
-                <meta charset="UTF-8" />
-                <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-                <title>Youtube API wrapper</title>
-              </head>
-              <body>
-                <div class="player-container">
-                  <div id="player"></div>
-                </div>
-              </body>
-            </html>
-        """.trimIndent()
-
         initJavascriptInterface()
         webView.setRendererPriorityPolicy(WebView.RENDERER_PRIORITY_IMPORTANT, false)
         webView.settings.javaScriptEnabled = true
         webView.settings.mediaPlaybackRequiresUserGesture = false
-        webView.webViewClient = object : WebViewClient() {
-            override fun shouldInterceptRequest(
-                view: WebView?,
-                request: WebResourceRequest?
-            ): WebResourceResponse? {
-                if (request?.url.toString().contains("ad.youtube.com") || request?.url.toString()
-                    .contains("ads.youtube.com")
-                ) {
-                    val textStream = ByteArrayInputStream("".toByteArray())
-                    return getTextWebResource(textStream)
-                }
-                return super.shouldInterceptRequest(view, request)
-            }
-        }
-        webView.loadDataWithBaseURL("https://www.youtube.com", iframe, "text/html", "utf-8", null)
-    }
-
-    private fun getTextWebResource(data: InputStream): WebResourceResponse? {
-        return WebResourceResponse("text/plain", "UTF-8", data)
+        webView.webViewClient = YoutubePlayerWebViewClient()
+        val displayMetrics = context.resources.displayMetrics
+        val dpHeight = displayMetrics.heightPixels / displayMetrics.density
+        webView.loadDataWithBaseURL(
+            "https://www.youtube.com",
+            youtubePlayerIframe(dpHeight.toInt()),
+            "text/html",
+            "utf-8",
+            null
+        )
     }
 
     private fun initJavascriptInterface() {
